Move over some files from common/ to core/
Left some files in common/, but this was kinda a judgment call.
