+++ core/test/LeveragedReturnCalculator.js
@@ -1,5 +1,5 @@
 const LeveragedReturnCalculator = artifacts.require("LeveragedReturnCalculator");
-const { didContractThrow } = require("../../common/SolidityTestUtils.js");
+const { didContractThrow } = require("./SolidityTestUtils.js");
 
 contract("LeveragedReturnCalculator", function(accounts) {
   // A deployed instance of the LeveragedReturnCalculator.
