+++ core/test/Registry.js
@@ -1,4 +1,4 @@
-const { didContractThrow } = require("../../common/SolidityTestUtils.js");
+const { didContractThrow } = require("./SolidityTestUtils.js");
 
 const truffleAssert = require("truffle-assertions");
